+++ models/utilities/constraint_check/src/constraint_set.cc
@@ -62,7 +62,7 @@ ConstraintSet::update()
   num_violations = 0;
   for (Constraint * constraint : constraints) {
     constraint->update();
-    num_violations += constraint->violated;
+    num_violations += static_cast<unsigned int>(constraint->violated);
   }
 }
